refactor(desktop): extract the shared invite-claim effect into useClaimInvite
PendingInviteGate and CommunityOnboardingFlow carried verbatim copies of
the claiming effect (same stage/error/isPending guard, claim call, stage
advance, error mapping, pending bookkeeping), kept from double-claiming
only by the mutual exclusivity of their mounts. Duplicated guards drift —
the transaction.error guard already had to be added twice in the same
shape.

Move the effect into a shared useClaimInvite hook in the onboarding
feature; both components now just call it. The hook owns its own pending
flag, which is equivalent for CommunityOnboardingFlow since its shared
isPending is only otherwise touched at the profile/team-intro stages,
never during claiming.

No behavior change: typecheck, biome, all desktop unit tests, and the six
deep-link e2e tests (gate claim, retry, connect variant, identity-swap
re-claim, no-double-claim, Joining handoff) pass.
